Test 4 testBuildSibling
Adding a sibling for an existing child implies that there is a way for
TagBuilder to know who the common parent is for the child and its new
sibling. There is currently no way to know this, as each TagNode instance
doesn’t store a reference to its parent. So I write the following failing test to drive the creation of this needed behavior