new performance logic:
The StringBuffer class is designed to automatically grow when needed. But there's a performance penalty because StringBuffer must work to transparently increase its size and shift data around. That performance penalty was not acceptable on the Evant system.

The solution lay in knowing the exact size the StringBuffer needs to be before instantiating it. How could we compute the appropriate size? Easy. As each node, attribute, or value is added to the TagBuilder, it can increment a buffer size based on the size of what is being added. The final computed buffer size can then be used to instantiate a StringBuffer that never needs to grow in size.